backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
aiomysql==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
